Move CommitVerificationConfig into CommitValidation
